+++ src/main/print.c
@@ -83,8 +83,8 @@ void PrintDefaults(void)
 {
     R_print.na_string = NA_STRING;
     R_print.na_string_noquote = mkChar("<NA>");
-    R_print.na_width = strlen(CHAR(R_print.na_string));
-    R_print.na_width_noquote = strlen(CHAR(R_print.na_string_noquote));
+    R_print.na_width = (int)strlen(CHAR(R_print.na_string));
+    R_print.na_width_noquote = (int)strlen(CHAR(R_print.na_string_noquote));
     R_print.quote = 1;
     R_print.right = Rprt_adj_left;
     R_print.digits = GetOptionDigits();
@@ -148,7 +148,7 @@ SEXP attribute_hidden do_prmatrix(SEXP call, SEXP op, SEXP args, SEXP rho)
         if (!isString(naprint) || LENGTH(naprint) < 1)
             error(_("invalid 'na.print' specification"));
         R_print.na_string = R_print.na_string_noquote = STRING_ELT(naprint, 0);
-        R_print.na_width = R_print.na_width_noquote = strlen(CHAR(R_print.na_string));
+        R_print.na_width = R_print.na_width_noquote = (int)strlen(CHAR(R_print.na_string));
     }
 
     if (length(rowlab) == 0)
@@ -256,7 +256,7 @@ SEXP attribute_hidden do_printdefault(SEXP call, SEXP op, SEXP args, SEXP rho)
         if (!isString(naprint) || LENGTH(naprint) < 1)
             error(_("invalid 'na.print' specification"));
         R_print.na_string = R_print.na_string_noquote = STRING_ELT(naprint, 0);
-        R_print.na_width = R_print.na_width_noquote = strlen(CHAR(R_print.na_string));
+        R_print.na_width = R_print.na_width_noquote = (int)strlen(CHAR(R_print.na_string));
     }
     args = CDR(args);
 
@@ -402,7 +402,7 @@ static void PrintGenericVector(SEXP s, SEXP env)
                 {
                     /* This can potentially overflow */
                     const char *ctmp = translateChar(STRING_ELT(tmp, 0));
-                    int len = strlen(ctmp);
+                    int len = (int)strlen(ctmp);
                     if (len < 100)
                         snprintf(pbuf, 115, "\"%s\"", ctmp);
                     else
@@ -452,7 +452,7 @@ static void PrintGenericVector(SEXP s, SEXP env)
     else
     { /* .. no dim() .. */
         names = getAttrib(s, R_NamesSymbol);
-        taglen = strlen(tagbuf);
+        taglen = (int)strlen(tagbuf);
         ptag = tagbuf + taglen;
         PROTECT(newcall = allocList(2));
         SETCAR(newcall, install("print"));
@@ -624,7 +624,7 @@ static void printList(SEXP s, SEXP env)
     else
     {
         i = 1;
-        taglen = strlen(tagbuf);
+        taglen = (int)strlen(tagbuf);
         ptag = tagbuf + taglen;
         PROTECT(newcall = allocList(2));
         SETCAR(newcall, install("print"));
@@ -1087,7 +1087,7 @@ attribute_hidden int F77_NAME(dblep0)(const char *label, int *nchar, double *dat
     int k, nc = *nchar;
 
     if (nc < 0)
-        nc = strlen(label);
+        nc = (int)strlen(label);
     if (nc > 255)
     {
         warning(_("invalid character length in dblepr"));
@@ -1109,7 +1109,7 @@ attribute_hidden int F77_NAME(intpr0)(const char *label, int *nchar, int *data, 
     int k, nc = *nchar;
 
     if (nc < 0)
-        nc = strlen(label);
+        nc = (int)strlen(label);
     if (nc > 255)
     {
         warning(_("invalid character length in intpr"));
@@ -1132,7 +1132,7 @@ attribute_hidden int F77_NAME(realp0)(const char *label, int *nchar, float *data
     double *ddata;
 
     if (nc < 0)
-        nc = strlen(label);
+        nc = (int)strlen(label);
     if (nc > 255)
     {
         warning(_("invalid character length in realpr"));
